+++ database/migrations/2025_02_21_150408_create_shopping_carts_table.php
@@ -11,11 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('shopping_carts', function (Blueprint $table) {
-            $table->id('shoppping_cart_id');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('shopping_carts')){
+            Schema::create('shopping_carts', function (Blueprint $table) {
+                $table->id('shoppping_cart_id');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
